requirements.txt
Django==4.2.6
pre-commit==3.4.0
django-environ==0.11.2
django-split-settings==1.2.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-filter==23.3
Markdown==3.5
djangorestframework-simplejwt==5.3.0
